antibody-mcts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
biopython==1.85
    # via
    #   -r requirements.in
    #   prodigy-prot
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
freesasa==2.2.1
    # via prodigy-prot
google-api-core[grpc]==2.24.2
    # via
    #   -r requirements.in
    #   google-cloud-core
    #   google-cloud-pubsub
    #   google-cloud-storage
google-auth==2.39.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-pubsub
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-pubsub==2.29.0
    # via -r requirements.in
google-cloud-storage==3.1.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-pubsub
grpcio==1.71.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
idna==3.10
    # via requests
importlib-metadata==8.6.1
    # via opentelemetry-api
numpy==2.2.4
    # via
    #   biopython
    #   prodigy-prot
opentelemetry-api==1.32.1
    # via
    #   google-cloud-pubsub
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.32.1
    # via google-cloud-pubsub
opentelemetry-semantic-conventions==0.53b1
    # via opentelemetry-sdk
prodigy-prot==2.2.5
    # via -r requirements.in
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
protobuf==5.29.4
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybktree==1.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9.1
    # via google-auth
typing-extensions==4.13.2
    # via opentelemetry-sdk
urllib3==2.4.0
    # via requests
wrapt==1.17.2
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
